add README.md

+++ micro-service-business/micro-service-module/src/main/java/com/winjean/service/impl/ModuleServiceImpl.java
@@ -15,6 +15,8 @@ import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.Optional;
+
 @Service
 @Slf4j
 @Transactional(readOnly = true)
@@ -37,9 +39,14 @@ public class ModuleServiceImpl implements ModuleService {
     @CacheEvict(cacheNames="ModuleCache",allEntries = true)
     //delete() 方法用于删除指定数据，由于delete方法删除的是实体，如果传递的参数不是实体那么需要先将待删除实体查询出来再进行删除
     public void delete(int id) {
-        ModuleEntity module =  moduleRepository.findById(id).get();
-        moduleRepository.delete(module);
-        log.info("module deleted id = {}",id);
+        Optional<ModuleEntity> optional =  moduleRepository.findById(id);
+        if (optional.isPresent()){
+            ModuleEntity module =  optional.get();
+            moduleRepository.delete(module);
+            log.info("module deleted id = {}",id);
+        }else{
+            log.info("no module exist, id = {}",id);
+        }
     }
 
     @Override
@@ -50,15 +57,21 @@ public class ModuleServiceImpl implements ModuleService {
     //update() 方法用于更新实体信息，spring data jpa对于数据的更新还有query update这种基于类sql的方式，这里我们不采用这种方式，我们采用的方式更简单update方法参数中包含了待更新实体的唯一标识，通过标识将原本的实体查询出来再将需要修改的属性值传递到查询中的实体就完成了数据库数据的更新十分简单
     public void update(ModuleEntity entity) {
         int id = entity.getId();
-        ModuleEntity module = moduleRepository.findById(id).get();
-        module.setName(entity.getName());
-        log.info("module updated id = {},name = {}", id, module.getName());
+        Optional<ModuleEntity> optional =  moduleRepository.findById(id);
+        if(optional.isPresent()){
+            ModuleEntity module = moduleRepository.findById(id).get();
+            module.setName(entity.getName());
+            log.info("module updated id = {},name = {}", id, module.getName());
+        }else{
+            log.info("no module exist, id = {}", id);
+        }
     }
 
     @Override
     @Cacheable(value = "ModuleCache", key = "#id")
     public ModuleEntity findById(int id) {
-        return moduleRepository.findById(id).get();
+        Optional<ModuleEntity> optional =  moduleRepository.findById(id);
+        return optional.isPresent() ? optional.get() : null;
     }
 
     @Override
